Extract NetSession (net + epoch guard) and renderRoomQR (#318 Step 3)
Increments 3a + 3b of the lobby decomposition — the contained, high-value
pieces of the join flow (the full MultiplayerJoinFlow class is deferred:
its ~12-point host interface is high-risk / modest-reward while the external
Lobby↔Game boundary is already clean).

3a — NetSession (js/lobby/net-session.js): owns the room NetworkManager
  instance + the stale-callback epoch guard. _replaceNet() delegates to
  session.replace() (tear down old net, epoch++, new net); the ~20 inline
  `netEpoch !== this._netEpoch` guards become `this._netSession.isCurrent(
  netEpoch)`. `this.net` is now a getter/setter over session.net, so every
  existing `this.net.*` read and `this.net = …` write is unchanged. Removes
  the now-unused NetworkManager import from lobby.js. Unit-tested: replace
  bumps epoch, old epochs go stale, manual null keeps the epoch (matching
  prior behavior).

3b — renderRoomQR (js/lobby/room-qr.js): the captain's join QR + shareable
  URL + long-press-copy block, lifted out of the 100-line onRoomJoined
  closure (joinUrl() also exported). Logic faithful; SITE_URL import moves
  with it.

Pure refactor — no behavior or wire change; the create/join/rejoin bodies
and their DOM/step coupling stay in Lobby. js/lobby/ now: room-store,
room-protocol, net-session, room-qr.
